took apart module helper methods into team,season,game
