+++ lib/stat_tracker.rb
@@ -7,6 +7,8 @@ require_relative '../modules/game_statables'
 require_relative '../modules/league_statables'
 require_relative '../modules/team_statables'
 require_relative '../modules/team_stat_helper'
+require_relative '../modules/season_stat_helper'
+require_relative '../modules/game_stat_helper'
 require_relative '../modules/season_statables'
 
 class StatTracker
@@ -15,6 +17,8 @@ class StatTracker
   include TeamStatables
   include SeasonStatables
   include TeamStatHelper
+  include SeasonStatHelper
+  include GameStatHelper
   attr_reader :games, :teams, :game_teams
 
   def initialize(games = {}, teams = {}, game_teams = {})
